# api/docs/swagger.yaml
basePath: /v1
definitions:
...
  entities.SubscriptionName:
    enum:
    - free
    - pro-monthly
    - pro-yearly
    - ultra-monthly
    - ultra-yearly
    - pro-lifetime
    - 20k-monthly
    - 100k-monthly
    - 50k-monthly
    - 20k-yearly
    - 50k-yearly
    - 100k-yearly
    - 200k-yearly
    - 200k-monthly
    type: string
    x-enum-varnames:
    - SubscriptionNameFree
    - SubscriptionNameProMonthly
    - SubscriptionNameProYearly
    - SubscriptionNameUltraMonthly
    - SubscriptionNameUltraYearly
    - SubscriptionNameProLifetime
    - SubscriptionName20KMonthly
    - SubscriptionName100KMonthly
    - SubscriptionName50KMonthly
    - SubscriptionName20KYearly
    - SubscriptionName50KYearly
    - SubscriptionName100KYearly
    - SubscriptionName200KYearly
    - SubscriptionName200KMonthly
  entities.User:
    properties:
      active_phone_id:
        example: 32343a19-da5e-4b1b-a767-3298a73703cb
        type: string
      api_key:
        example: x-api-key
        type: string
      created_at:
        example: "2022-06-05T14:26:02.302718+03:00"
        type: string
      email:
        example: name@email.com
        type: string
      id:
        example: WB7DRDWrJZRGbYrv2CKGkqbzvqdC
        type: string
      is_admin:
        example: true
        type: boolean
      locale:
        example: pt-BR
        type: string
      notification_heartbeat_enabled:
        example: true
        type: boolean
      notification_message_status_enabled:
        example: true
        type: boolean
      notification_newsletter_enabled:
        example: true
        type: boolean
      notification_webhook_enabled:
        example: true
        type: boolean
      subscription_ends_at:
        example: "2022-06-05T14:26:02.302718+03:00"
        type: string
      subscription_id:
        example: 8f9c71b8-b84e-4417-8408-a62274f65a08
        type: string
      subscription_name:
        allOf:
        - $ref: '#/definitions/entities.SubscriptionName'
        example: free
      subscription_renews_at:
        example: "2022-06-05T14:26:02.302718+03:00"
        type: string
      subscription_status:
        example: on_trial
        type: string
      timezone:
        example: Europe/Helsinki
        type: string
      updated_at:
        example: "2022-06-05T14:26:10.303278+03:00"
        type: string
    required:
    - api_key
    - created_at
    - email
    - id
    - is_admin
    - locale
    - notification_heartbeat_enabled
    - notification_message_status_enabled
    - notification_newsletter_enabled
    - notification_webhook_enabled
    - subscription_id
    - subscription_name
    - timezone
    - updated_at
    type: object
...
  responses.Unauthorized:
    properties:
      data:
        example: Make sure your API key is set in the [X-API-Key] header in the request
        type: string
      message:
        example: You are not authorized to carry out this request.
        type: string
      status:
        example: error
        type: string
    required:
    - data
    - message
    - status
    type: object
...
  responses.UserListResponse:
    properties:
      data:
        properties:
          items:
            items:
              $ref: '#/definitions/entities.User'
            type: array
          total_count:
            type: integer
        required:
        - items
        - total_count
        type: object
      message:
        example: Request handled successfully
        type: string
      status:
        example: success
        type: string
    required:
    - data
    - message
    - status
    type: object
...
host: api.httpsms.com
info:
  contact:
    email: support@httpsms.com
    name: support@httpsms.com
  description: Use your Android phone to send and receive SMS messages via a simple
    programmable API with end-to-end encryption.
  license:
    name: AGPL-3.0
    url: https://raw.githubusercontent.com/NdoleStudio/http-sms-manager/main/LICENSE
  title: httpSMS API Reference
  version: "1.0"
paths:
...
  /v1/admin/users:
    get:
      consumes:
      - application/json
      description: Fetches a paginated list of all users.
      parameters:
      - default: 0
        description: Skip
        in: query
        name: skip
        type: integer
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - description: Search by email
        in: query
        name: query
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.UserListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Unauthorized'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.Unauthorized'
      security:
      - BearerAuth: []
      summary: Fetch all users (Admin only)
      tags:
      - Admin
...
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-api-Key
    type: apiKey
swagger: "2.0"